Fix Transaction Job

+++ resources/views/livewire/purchasing/transaction/purchase-order/detail.blade.php
@@ -1,10 +1,18 @@
 <div>
     <form wire:submit="store">
         <div class='row'>
+            {{-- NUMBER --}}
+            @if ($number)
+                <div class="col-md-4 mb-3">
+                    <label>Nomor</label>
+                    <input type="text" class='form-control' value="{{ $number }}" disabled>
+                </div>
+            @endif
+
             {{-- SELECT COMPANY --}}
             <div class="col-md-4 mb-3 {{ $isMultipleCompany ? '' : 'd-none' }}">
                 <label>Perusahaan</label>
-                <select class="form-select w-100" wire:model='companyId'>
+                <select class="form-select w-100" wire:model='companyId' {{ $isShow ? 'disabled' : '' }}>
                     @php $isFound = false; @endphp
 
                     @foreach ($companies as $company)
@@ -21,7 +29,7 @@
             {{-- SELECT WAREHOUSE --}}
             <div class="col-md-4 mb-3">
                 <label>Gudang</label>
-                <select class="form-select w-100" wire:model='warehouseId'>
+                <select class="form-select w-100" wire:model='warehouseId' {{ $isShow ? 'disabled' : '' }}>
                     @php $isFound = false; @endphp
 
                     @foreach ($warehouses as $warehouse)
@@ -38,7 +46,7 @@
             {{-- SELECT SUPPLIER --}}
             <div class="col-md-4 mb-3" wire:ignore>
                 <label>Supplier</label>
-                <select id="select2-supplier" class="form-select w-100">
+                <select id="select2-supplier" class="form-select w-100" {{ $isShow ? 'disabled' : '' }}>
                     @if ($supplierId)
                         <option value="{{ $supplierId }}">{{ $supplierText }}</option>
                     @endif
@@ -49,7 +57,7 @@
             <div class="col-md-4 mb-3">
                 <label>Tanggal</label>
                 <input type="date" class="form-control @error('transactionDate') is-invalid @enderror"
-                    wire:model="transactionDate" />
+                    wire:model="transactionDate" {{ $isShow ? 'disabled' : '' }} />
 
                 @error('transactionDate')
                     <div class="invalid-feedback">
@@ -62,7 +70,7 @@
             <div class="col-md-4 mb-3">
                 <label>Nomor Nota Supplier</label>
                 <input type="text" class="form-control @error('supplierInvoiceNumber') is-invalid @enderror"
-                    wire:model="supplierInvoiceNumber" />
+                    wire:model="supplierInvoiceNumber" {{ $isShow ? 'disabled' : '' }} />
 
                 @error('supplierInvoiceNumber')
                     <div class="invalid-feedback">
@@ -74,13 +82,13 @@
             {{-- NOTE --}}
             <div class="col-md-12 mb-3">
                 <label>Catatan</label>
-                <textarea class="form-control" cols="30" rows="4" wire:model="note"></textarea>
+                <textarea class="form-control" cols="30" rows="4" wire:model="note" {{ $isShow ? 'disabled' : '' }}></textarea>
             </div>
         </div>
 
         {{-- PRODUCTS --}}
         <label>Barang-barang yang diterima</label>
-        <div class="col-md-12 mb-4" wire:ignore>
+        <div class="col-md-12 mb-4 {{ $isShow ? 'd-none' : '' }}" wire:ignore>
             <select id="select2-product" class="form-select w-100">
             </select>
         </div>
@@ -104,30 +112,32 @@
                 <tr>
                     {{-- ACTION --}}
                     <td style="width: 2%" class='align-bottom'>
-                        <label class='fw-bold'>Aksi</label>
-                        <button type="button"
-                            class="btn btn-outline btn-outline-dashed btn-outline-secondary btn-active-light-secondary dropdown-toggle dropdown-toggle-split"
-                            data-bs-toggle="dropdown" aria-expanded="false">
-                            <span class="visually-hidden">Toggle Dropdown</span>
-                        </button>
-                        <ul class="dropdown-menu">
-                            <li>
-                                <button type="button" class="dropdown-item text-info"
-                                    wire:click.prevent="duplicateDetail({{ $index }})">
-                                    <i class="ki-solid ki-copy text-info"></i>
-                                    Duplikat
-                                </button>
-                            </li>
-                            @if ($item['is_deletable'])
+                        @if (!$isShow)
+                            <label class='fw-bold'>Aksi</label>
+                            <button type="button"
+                                class="btn btn-outline btn-outline-dashed btn-outline-secondary btn-active-light-secondary dropdown-toggle dropdown-toggle-split"
+                                data-bs-toggle="dropdown" aria-expanded="false">
+                                <span class="visually-hidden">Toggle Dropdown</span>
+                            </button>
+                            <ul class="dropdown-menu">
                                 <li>
-                                    <button type="button" class="dropdown-item text-danger"
-                                        wire:click="removeDetail({{ $index }})">
-                                        <i class="ki-solid ki-abstract-11 text-danger"></i>
-                                        Hapus
+                                    <button type="button" class="dropdown-item text-info"
+                                        wire:click.prevent="duplicateDetail({{ $index }})">
+                                        <i class="ki-solid ki-copy text-info"></i>
+                                        Duplikat
                                     </button>
                                 </li>
-                            @endif
-                        </ul>
+                                @if ($item['is_deletable'])
+                                    <li>
+                                        <button type="button" class="dropdown-item text-danger"
+                                            wire:click="removeDetail({{ $index }})">
+                                            <i class="ki-solid ki-abstract-11 text-danger"></i>
+                                            Hapus
+                                        </button>
+                                    </li>
+                                @endif
+                            </ul>
+                        @endif
                     </td>
 
                     {{-- NAME --}}
@@ -141,10 +151,12 @@
                         <label class='fw-bold'>Jumlah</label>
                         <div class="input-group">
                             <input type="text" class="form-control currency"
-                                wire:model.blur="purchaseOrderProducts.{{ $index }}.quantity" />
+                                wire:model.blur="purchaseOrderProducts.{{ $index }}.quantity"
+                                {{ $isShow ? 'disabled' : '' }} />
 
                             <select class="form-select @error('type') is-invalid @enderror"
-                                wire:model.blur="purchaseOrderProducts.{{ $index }}.unit_detail_id">
+                                wire:model.blur="purchaseOrderProducts.{{ $index }}.unit_detail_id"
+                                {{ $isShow ? 'disabled' : '' }}>
                                 @foreach ($item['unit_detail_choice'] as $unit)
                                     <option value="{{ $unit['id'] }}">
                                         {{ $unit['name'] }}
@@ -159,7 +171,8 @@
                     <td style="width: 25%;">
                         <label class='fw-bold'>Harga Satuan</label>
                         <input type="text" class="form-control currency"
-                            wire:model.blur="purchaseOrderProducts.{{ $index }}.price" />
+                            wire:model.blur="purchaseOrderProducts.{{ $index }}.price"
+                            {{ $isShow ? 'disabled' : '' }} />
                     </td>
 
                     {{-- TAX --}}
@@ -167,7 +180,8 @@
                         <label class='fw-bold'>Pajak</label>
                         <div class="form-check mt-3">
                             <input class="form-check-input" type="checkbox" value="" id="ppn_{{ $index }}"
-                                wire:model.live="purchaseOrderProducts.{{ $index }}.is_ppn">
+                                wire:model.live="purchaseOrderProducts.{{ $index }}.is_ppn"
+                                {{ $isShow ? 'disabled' : '' }}>
                             <label class="form-check-label" for="ppn_{{ $index }}">
                                 PPN
                             </label>
@@ -191,7 +205,8 @@
                             <td>
                                 <label class='fw-bold'>Kode Barang</label>
                                 <input type="text" class="form-control"
-                                    wire:model="purchaseOrderProducts.{{ $index }}.code" />
+                                    wire:model="purchaseOrderProducts.{{ $index }}.code"
+                                    {{ $isShow ? 'disabled' : '' }} />
                             </td>
                         @endif
 
@@ -200,7 +215,8 @@
                             <td>
                                 <label class='fw-bold'>Kode Produksi</label>
                                 <input type="text" class="form-control"
-                                    wire:model="purchaseOrderProducts.{{ $index }}.batch" />
+                                    wire:model="purchaseOrderProducts.{{ $index }}.batch"
+                                    {{ $isShow ? 'disabled' : '' }} />
                             </td>
                         @endif
 
@@ -210,7 +226,8 @@
                                 <label class='fw-bold'>Tanggal Expired</label>
                                 <input type="date"
                                     class="form-control @error('purchaseOrderProducts.{{ $index }}.expired_date') is-invalid @enderror"
-                                    wire:model="purchaseOrderProducts.{{ $index }}.expired_date" />
+                                    wire:model="purchaseOrderProducts.{{ $index }}.expired_date"
+                                    {{ $isShow ? 'disabled' : '' }} />
                                 @error('purchaseOrderProducts.{{ $index }}.expired_date')
                                     <div class="invalid-feedback">
                                         {{ $message }}
@@ -227,28 +244,32 @@
                         <td colspan="5">
                             {{-- ATTACHMENTS --}}
                             <label class='fw-bold'>Lampiran</label>
-                            <input id="fileInput_{{ $index }}" class="form-control d-none" type="file"
-                                wire:model="purchaseOrderProducts.{{ $index }}.files" multiple>
 
-                            <button type="button" class="btn btn-info btn-sm"
-                                onclick="$('#fileInput_{{ $index }}').click()" wire:loading.attr="disabled"
-                                wire:target='purchaseOrderProducts.{{ $index }}.files'>
-                                <div wire:loading wire:target='purchaseOrderProducts.{{ $index }}.files'>
-                                    <span class="spinner-border spinner-border-sm" role="status"
-                                        aria-hidden="true"></span>
-                                    Loading...
-                                </div>
+                            @if (!$isShow)
+                                <input id="fileInput_{{ $index }}" class="form-control d-none" type="file"
+                                    wire:model="purchaseOrderProducts.{{ $index }}.files" multiple>
 
-                                <div wire:loading.class="d-none"
+                                <button type="button" class="btn btn-info btn-sm"
+                                    onclick="$('#fileInput_{{ $index }}').click()"
+                                    wire:loading.attr="disabled"
                                     wire:target='purchaseOrderProducts.{{ $index }}.files'>
-                                    <i class="ki-solid ki-double-up"></i>
-                                    Tambah Lampiran
-                                </div>
-                            </button>
+                                    <div wire:loading wire:target='purchaseOrderProducts.{{ $index }}.files'>
+                                        <span class="spinner-border spinner-border-sm" role="status"
+                                            aria-hidden="true"></span>
+                                        Loading...
+                                    </div>
 
-                            @error('files.*')
-                                <span class="error">{{ $message }}</span>
-                            @enderror
+                                    <div wire:loading.class="d-none"
+                                        wire:target='purchaseOrderProducts.{{ $index }}.files'>
+                                        <i class="ki-solid ki-double-up"></i>
+                                        Tambah Lampiran
+                                    </div>
+                                </button>
+
+                                @error('files.*')
+                                    <span class="error">{{ $message }}</span>
+                                @enderror
+                            @endif
 
                             <div class="mt-3">
                                 @if ($purchaseOrderProducts[$index]['uploadedFiles'])
@@ -262,7 +283,8 @@
                                             <div class="col-md-6 mb-2">
                                                 <input type="text" class="form-control form-control-sm"
                                                     cols="30" rows="4"
-                                                    wire:model="purchaseOrderProducts.{{ $index }}.uploadedFiles.{{ $fileIndex }}.note" />
+                                                    wire:model="purchaseOrderProducts.{{ $index }}.uploadedFiles.{{ $fileIndex }}.note"
+                                                    {{ $isShow ? 'disabled' : '' }} />
                                             </div>
                                             <div class="col-md-6 mb-2">
                                                 <a href="{{ $file['url'] }}" target="_blank"
@@ -270,10 +292,12 @@
                                                     <i class="ki-solid ki-file"></i>
                                                     {{ $file['original_file_name'] }}
                                                 </a>
-                                                <button type="button" class="btn btn-icon btn-danger btn-sm"
-                                                    wire:click="removeFile('{{ $index }}', '{{ $fileIndex }}')">
-                                                    <i class="ki-solid ki-abstract-11"></i>
-                                                </button>
+                                                @if (!$isShow)
+                                                    <button type="button" class="btn btn-icon btn-danger btn-sm"
+                                                        wire:click="removeFile('{{ $index }}', '{{ $fileIndex }}')">
+                                                        <i class="ki-solid ki-abstract-11"></i>
+                                                    </button>
+                                                @endif
                                             </div>
                                             <hr class="d-md-none">
                                         @endforeach
@@ -321,31 +345,17 @@
             @endif
         </table>
 
-        <button type="submit" class="btn btn-success mt-3">
-            <i class='ki-duotone ki-check fs-1'></i>
-            Simpan
-        </button>
+        @if (!$isShow)
+            <button type="submit" class="btn btn-success mt-3">
+                <i class='ki-duotone ki-check fs-1'></i>
+                Simpan
+            </button>
+        @endif
     </form>
 
     {{-- HISTORY DATATABLE --}}
     @if ($objId)
-        <div class="accordion mt-4" id="accordionExample" wire:ignore.self>
-            <div class="accordion-item" wire:ignore.self>
-                <h2 class="accordion-header" id="headingOne" wire:ignore>
-                    <button class="accordion-button" type="button" data-bs-toggle="collapse"
-                        data-bs-target="#collapseOne" aria-expanded="true" aria-controls="collapseOne">
-                        <label class='fw-bold'>Kartu Stok</label>
-                    </button>
-                </h2>
-                <div id="collapseOne" class="accordion-collapse collapse show" aria-labelledby="headingOne"
-                    data-bs-parent="#accordionExample" wire:ignore.self>
-                    <div class="accordion-body" wire:ignore.self>
-                        <livewire:logistic.transaction.product-detail-history.history-datatable :remarksIds="$historyRemarksIds"
-                            :remarksType="$historyRemarksType" />
-                    </div>
-                </div>
-            </div>
-        </div>
+        <livewire:logistic.transaction.product-detail-history.history-datatable :remarksIds="$historyRemarksIds" :remarksType="$historyRemarksType" />
     @endif
 </div>
 @push('css')
